requirements.txt
openai==1.12.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
weaviate-client==3.24.2
numpy>=1.26.4
pandas>=2.2.3
